[FEATURE] Initial journal decorator.
Adds a journal decorator data structure which allows to store delta events such as insertion, deletion and substitution in a journal.

The data structure uses a dictionary with logarithmic access time to the entires.

A journal entry stores the target sequence position and the associated infix.
This is currently implemented with a span.
The journal decorator does not own its memory, except the journal dictionary.

The journal decorator offers a dedicated sequence interface though.
It implements a random access iterator with
logarithmic time access.