[FEATURE] Prepare journal sequence tree to convert alignments into delta events.
This change transforms an alignment into a set of delta events and stores the respective banch and join events in sorted event queues.
These data structures store only a fraction of the original sequence content and allow a parallel traversal
over all sequences with a given sequence context.
The additional function to get the context enumerator is suffixed with _2 and will replace the old interface, once everything works as expected.