[FEATURE] Initial vcf parser.
This feature parses a vcf file using SeqAn2 vcf I/O.
It currently supports SNP data inside of the vcf.
It generates for every vcf record a list of shared delta events which is added ot the JST.
The constructed JST is then returned.
In the test the consistency is checked by comparing the JST sequences against the
set of haplotype sequences that are simulated together with the vcf file.