[MISC] Make jst move only.
The JST stores internally a set of pointers to internally managed nodes.
When copying the pointer must be carefully reassigned to the copied nodes.
In general, there is little need in copying a JST rather than copy the agents working on the JST.
The jst is kind of a large index, where copying might be expensive.
To avoid subtle performance bugs, when the jst is copied rather than moved or passed by reference the copyable property is removed.