[FIX] Use the default copy constructor.
The version before had a negative performance impact. When copying a std::vector there is no guarantee that
the new allocated memory has the same capacity as the source as long as all elements fit into the new cpacity.
Now the bit vector always allocates sufficiently large memory which is not addressed by the operations.