[MISC] Remove unused parts of the jst traversal.
This a lot of the traversal which was based on the frequent update of the base coverage.
This is not needed as the positions can be retrieved in a separate step.
During the traversal most algorithms would not require to get all sequence positions.
Also the offsets for each bin in the partitioned jst are not needed to be computed.