[MISC] Refactor branch to store subtree root.
Before we were using the current delta event to create a new branch.
But most operations are only used inside of the branch creation and only need the current branch event as an additional parameter.
Instead we store the root of the subtree inside of the branch.
Using this root we do not need to access the stack anymore at a specific position.