[MISC] Refactor filter step to output bins containing query ids.
The reported bins from the filter step now return query ids instead of sequences.
First, within the thread the ids are transfromed to ref views over the queries.
In addition the local ids in the reported matches are remapped to the global query id.
Thus, the output function can work with the entire query set instead of batches.