Introduced a new concept `preserving_reference_sequence` in `reference_sequence_concept.hpp.`
This concept checks if an object is a reference sequence and if its breakpoint slice is also a reference sequence. It also ensures that the nested breakpoint slice can be converted to the breakpoint slice of the object, and that the sequence breakpoint of the breakpoint slice of the object can be converted to the sequence breakpoint of the object.
This guarantees that the CPOs `breakpoint_slice` and `to_breakpoint` can be invoked recursively on the breakpoint_slice returned from the original object.